This commit changed and added functionality to the database layer. Hibernate persistence used instead of JPA EntityManager. Repository DAO classes modified.
Test cases added for Database Layer

+++ version-review-test-client/src/com/heinvdende/versionreview/test/modules/repository/persistence/CodeFileRepository.java
@@ -12,7 +12,7 @@ import com.heinvdende.versionreview.test.modules.repository.domain.CodeFile;
  *
  * @author Heinrich
  */
-public class CodeFileRepository extends AbstractRepository<CodeFile, Long> {
+public class CodeFileRepository extends AbstractRepository<CodeFile, Integer> {
 
     @Override
     public Class getEntityClass() {
