+++ version-review-test-client/src/com/heinvdende/versionreview/test/modules/repository/persistence/MemberRepository.java
@@ -12,7 +12,7 @@ import com.heinvdende.versionreview.test.modules.repository.domain.ClassMember;
  *
  * @author Heinrich
  */
-public class MemberRepository extends AbstractRepository<ClassMember, Long> {
+public class MemberRepository extends AbstractRepository<ClassMember, Integer> {
 
     @Override
     public Class getEntityClass() {
