+++ version-review-test-client/src/com/heinvdende/versionreview/test/modules/repository/persistence/TaskClassRepository.java
@@ -12,7 +12,7 @@ import com.heinvdende.versionreview.test.modules.repository.domain.TaskClass;
  *
  * @author Heinrich
  */
-public class TaskClassRepository extends AbstractRepository<TaskClass, Long> {
+public class TaskClassRepository extends AbstractRepository<TaskClass, Integer> {
 
     @Override
     public Class getEntityClass() {
